+++ lib/src/models/pick_result.dart
@@ -5,22 +5,58 @@ class PickResult {
   PickResult({
     this.placeId,
     this.geometry,
-    this.address,
+    this.formattedAddress,
     this.types,
     this.addressComponents,
+    this.adrAddress,
+    this.formattedPhoneNumber,
+    this.id,
+    this.reference,
+    this.icon,
+    this.name,
+    this.openingHours,
+    this.photos,
+    this.internationalPhoneNumber,
+    this.priceLevel,
+    this.rating,
+    this.scope,
+    this.url,
+    this.vicinity,
+    this.utcOffset,
+    this.website,
+    this.reviews,
   });
 
   final String placeId;
   final Geometry geometry;
-  final String address;
+  final String formattedAddress;
   final List<String> types;
   final List<AddressComponent> addressComponents;
+
+  // Below results will not be fetched if 'usePlaceDetailSearch' is set to false (Defaults to false).
+  final String adrAddress;
+  final String formattedPhoneNumber;
+  final String id;
+  final String reference;
+  final String icon;
+  final String name;
+  final OpeningHoursDetail openingHours;
+  final List<Photo> photos;
+  final String internationalPhoneNumber;
+  final PriceLevel priceLevel;
+  final num rating;
+  final String scope;
+  final String url;
+  final String vicinity;
+  final num utcOffset;
+  final String website;
+  final List<Review> reviews;
 
   factory PickResult.fromGeocodingResult(GeocodingResult result) {
     return PickResult(
       placeId: result.placeId,
       geometry: result.geometry,
-      address: result.formattedAddress,
+      formattedAddress: result.formattedAddress,
       types: result.types,
       addressComponents: result.addressComponents,
     );
@@ -30,9 +66,26 @@ class PickResult {
     return PickResult(
       placeId: result.placeId,
       geometry: result.geometry,
-      address: result.formattedAddress,
+      formattedAddress: result.formattedAddress,
       types: result.types,
       addressComponents: result.addressComponents,
+      adrAddress: result.adrAddress,
+      formattedPhoneNumber: result.formattedPhoneNumber,
+      id: result.id,
+      reference: result.reference,
+      icon: result.icon,
+      name: result.name,
+      openingHours: result.openingHours,
+      photos: result.photos,
+      internationalPhoneNumber: result.internationalPhoneNumber,
+      priceLevel: result.priceLevel,
+      rating: result.rating,
+      scope: result.scope,
+      url: result.url,
+      vicinity: result.vicinity,
+      utcOffset: result.utcOffset,
+      website: result.website,
+      reviews: result.reviews,
     );
   }
 }